requirement.txt
torchaudio
scipy
matplotlib
pillow
librosa
numpy<2.0